change files to js to avoid build issues when using grafana docker container
