+++ scripts/K6/src/no-access-with-certs-smoke-tests.js
@@ -1,11 +1,10 @@
-/// <reference types="k6" />
-import http from 'k6/http';
-import { check } from 'k6';
+const http = require('k6/http');
+const { check } = require('k6');
 
 const baseUrl = 'https://dev.integration-api.hmpps.service.justice.gov.uk';
 const notAllowedEndpoint = `/v1/persons?first_name=john`;
 
-export const options = {
+module.exports.options = {
   tlsAuth: [
     {
       domains: ['dev.integration-api.hmpps.service.justice.gov.uk'],
@@ -15,7 +14,7 @@ export const options = {
   ],
 };
 
-export default function () {
+module.exports.default = function () {
   const res = http.get(`${baseUrl}${notAllowedEndpoint}`, {
     headers: {
       'x-api-key': __ENV.NO_ACCESS_API_KEY,
@@ -25,4 +24,4 @@ export default function () {
   check(res, {
     'Cert-based no-access returns 403': (r) => r.status === 403,
   });
-}
+};
